backend/pyproject.toml
[tool.poetry]
name = "payment-service"
version = "0.1.0"
authors = ["Anton Kimyaev"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
djangorestframework = "^3.16.1"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.1"
celery = "^5.3.4"
redis = "^5.0.1"
ruff = "^0.11.4"
drf-spectacular = "^0.29.0"
pytest = "^8.0.0"
factory-boy = "^3.3.0"
pytest-django = "^4.8.0"
pytest-mock = "^3.12.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
src = ["."]
target-version = "py312"

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "__init__.py"
]

line-length = 100

[tool.ruff.format]
quote-style = "double"
docstring-code-format = true
docstring-code-line-length = 100

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "B", "Q", "I"]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
ignore = [
    "RUF001",
    "S101",
    "TID252",
    "D102",
    "D100",
    "RUF002",
    "ANN001",
    "ANN201",
    "A003",
    "RUF012",
    "D106",
    "D101",
    "W191",
    "ERA001",
    "ANN002",
    "ANN003",
    "Q000",
    "Q003",
    "D206",
    "COM812",
    "D203",
    "E731",
    "B904",
    "B007",
    "E712",
    "E711",
    "E501",
    "B008",
    "B006",
    "B020",
    "B018",
    "F722",
    "F405",
    "F403"
]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.pycodestyle]
max-doc-length = 100
max-line-length = 100
